handled get and post users routes
