+++ models/userModel.js
@@ -18,6 +18,6 @@ const UserSchema = new Schema({
     }
 });
 
-const User = mongoose.model('User', UserSchema);
+const User = mongoose.model('user', UserSchema);
 
 module.exports = User;
