+++ package.json
@@ -11,12 +11,15 @@
   "license": "ISC",
   "dependencies": {
     "bcrypt": "^5.1.0",
+    "cookie-parser": "^1.4.6",
     "dotenv": "^16.0.3",
     "express": "^4.18.1",
+    "jsonwebtoken": "^8.5.1",
     "mocha": "^10.0.0",
     "moment": "^2.29.4",
     "mongoose": "^6.6.1",
-    "save": "^2.9.0"
+    "save": "^2.9.0",
+    "validator": "^13.7.0"
   },
   "devDependencies": {
     "nodemon": "^2.0.20"
